Move the PayrollDataCsv into a new Payroll module
